MaryiaF/refactor: convert CFDPOA component to ts
