+++ packages/trader/src/Modules/CFD/Containers/cfd-financial-stp-real-account-signup.jsx
@@ -4,7 +4,7 @@ import { FormProgress, DesktopWrapper, MobileWrapper, Div100vhContainer } from '
 import { getPropertyValue, isDesktop, WS } from '@deriv/shared';
 import { localize, Localize } from '@deriv/translations';
 import { connect } from 'Stores/connect';
-import CFDPOA from '../Components/cfd-poa.jsx';
+import CFDPOA from '../Components/cfd-poa';
 import CFDPersonalDetailsForm from '../Components/cfd-personal-details-form';
 import CFDPOI from '../Components/cfd-poi';
